adding ownership protection

+++ src/features/ticket/actions/delete-ticket.ts
@@ -1,16 +1,28 @@
 "use server";
+import {isOwner} from "../../../features/auth/utils/is-owner";
 import {setCookieByKey} from "../../../actions/cookies";
 import {revalidatePath} from "next/cache";
-import {fromErrorToActionState} from "../../../components/form/utils/to-action-state";
+import {fromErrorToActionState, toActionState} from "../../../components/form/utils/to-action-state";
+import {getAuthOrRedirect} from "../../../features/auth/queries/get-auth-or-redirect";
 import {redirect} from "next/navigation";
 import prisma from "../../../lib/prisma";
 import {ticketsPath} from "../../../app/paths";
 
 export const deleteTicket = async (id: string) =>
 {   
-    await new Promise((resolve) => setTimeout(resolve, 2000));
+    const {user} = await getAuthOrRedirect();
 
     try {
+        const ticket = await prisma.ticket.findUnique({
+            where: {
+                id,
+            },
+        });
+
+        if (!ticket || !isOwner(user, ticket)) {
+            return toActionState("ERROR", "Not authorized");
+        }
+
         await prisma.ticket.delete({
             where: {
                 id,
